refactor: POM Super Class 분리
