+++ client/tests/POM/pom.ts
@@ -1,0 +1,9 @@
+import { Page } from "@playwright/test";
+
+export default class POM {
+  readonly page: Page;
+
+  constructor(page: Page) {
+    this.page = page;
+  }
+}
